Split up PDMA into modules
